# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  models.Connection:
    properties:
      createdAt:
        type: string
      database:
        type: string
      host:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      options:
        additionalProperties: true
        description: Database-specific options (warehouse, role, schema, etc)
        type: object
      password:
        description: AES-256 Encrypted
        type: string
      port:
        type: integer
      type:
        description: postgres, mysql, bigquery, etc
        type: string
      updatedAt:
        type: string
      userId:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@insightengine.ai
    name: API Support
  description: Enterprise BI Platform API with Power BI/Tableau parity.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: InsightEngine API
  version: "1.0"
paths:
...
  /connection:
    get:
      consumes:
      - application/json
      description: Returns a list of database connections for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List connections
      tags:
      - Connection
    post:
      consumes:
      - application/json
      description: Creates a new database connection.
      parameters:
      - description: Connection Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Connection'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a connection
      tags:
      - Connection
  /connection/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a database connection by ID.
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a connection
      tags:
      - Connection
    get:
      consumes:
      - application/json
      description: Returns a single database connection by ID.
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a connection
      tags:
      - Connection
    put:
      consumes:
      - application/json
      description: Updates an existing database connection.
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: string
      - description: Connection Updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Connection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a connection
      tags:
      - Connection
  /connection/{id}/schema:
    get:
      consumes:
      - application/json
      description: Discovers and returns the schema for a database connection.
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get connection schema
      tags:
      - Connection
  /connection/{id}/test:
    post:
      consumes:
      - application/json
      description: Tests connectivity to a database connection.
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Test connection
      tags:
      - Connection
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"